add long

+++ src/Main.java
@@ -1,8 +1,9 @@
 public class Main {
     public static void main(String[] args) {
-        int[] sales = new int[] {1, 5, 600};
+        long[] sales = new long[] {1, 5, 600};
         SalesManager salesManager = new SalesManager(sales);
-        int max = salesManager.max();
+
+        long max = salesManager.max();
         System.out.println(max);
     }
 }
